fix: add option to use builtin SSH client, not enabled by default
On some windows machines the builtin SSH library doesn't play well with DevPod. Originally we've tried to fix this by using our own SSH client. A couple of users reported this change broke their setup. We're rolling back the default on windows and instead offer it as a workaround via the `USE_BUILTIN_SSH` option. Updated readme for troubleshooting